add auth token with jwt to nest js. update function get list, detail cashier, refactor cashier controller

+++ src/cashier/cashier.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { JwtService } from '@nestjs/jwt';
 import { InjectRepository } from '@nestjs/typeorm';
-import { processResponse } from 'src/shared/common';
+import PayloadToken from 'src/auth/interface/payloadToken.interface';
 import { Repository } from 'typeorm';
 import { CreateCashierDto } from './dto/create-cashier.dto';
 import { FilterCashierDto } from './dto/filter-cashier.dto';
@@ -9,17 +10,34 @@ import { Cashier } from './entities/cashier.entity';
 
 @Injectable()
 export class CashierService {
-  constructor(@InjectRepository(Cashier) private cashierRepository: Repository<Cashier>) {}
+  constructor(
+    @InjectRepository(Cashier) private cashierRepository: Repository<Cashier>,
+    private jwtService: JwtService,
+  ) {}
 
   async create(createCashierDto: CreateCashierDto) {
     const createCategory = await this.cashierRepository.save(createCashierDto);
-    return processResponse(true, createCategory);
+    return createCategory;
+  }
+
+  getToken(cashier: Cashier) {
+    const payload: PayloadToken = { passcode: cashier.passcode, CashierId: cashier.cashierId };
+    const token = this.jwtService.sign(payload);
+    return token;
+  }
+
+  async validate(cashierId, passcode) {
+    const cashier = await this.cashierRepository.findOne({
+      where: { cashierId, passcode },
+      select: ['passcode', 'cashierId', 'passcode'],
+    });
+    return cashier;
   }
 
   async findAll(query: FilterCashierDto) {
     const { limit = 10, skip = 0 } = query;
 
-    const cashierList: CreateCashierDto[] = await this.cashierRepository.find({
+    const [cashierList, total] = await this.cashierRepository.findAndCount({
       select: ['cashierId', 'name'],
       take: limit,
       skip: skip,
@@ -28,29 +46,54 @@ export class CashierService {
     const data = {
       cashiers: cashierList,
       meta: {
-        total: cashierList.length,
+        total,
         limit,
         skip,
       },
     };
-    return processResponse(true, data);
+    return data;
   }
 
   async findOne(id: number) {
     const cashier = await this.cashierRepository.findOne({
       where: { cashierId: id },
-      select: ['cashierId', 'name']
+      select: ['cashierId', 'name'],
     });
-    return processResponse(true, cashier);
+
+    if (!cashier) throw new NotFoundException('Cashier not found');
+
+    return cashier;
+  }
+
+  async findPasscodeById(id: number) {
+    const cashier = await this.cashierRepository.findOne({
+      where: { cashierId: id },
+      select: ['passcode'],
+    });
+
+    if (!cashier) throw new NotFoundException('Cashier not found');
+
+    return { passcode: cashier.passcode };
   }
 
   async update(id: number, updateCashierDto: UpdateCashierDto) {
+    const cashier = await this.cashierRepository.findOne({
+      where: { cashierId: id },
+      select: ['cashierId', 'name'],
+    });
+
+    if (!cashier) throw new NotFoundException('Cashier not found');
     await this.cashierRepository.update({ cashierId: id }, updateCashierDto);
-    return processResponse(true);
   }
 
   async remove(id: number) {
+    const cashier = await this.cashierRepository.findOne({
+      where: { cashierId: id },
+      select: ['cashierId', 'name'],
+    });
+
+    if (!cashier) throw new NotFoundException('Cashier not found');
+
     await this.cashierRepository.delete({ cashierId: id });
-    return processResponse(true);
   }
 }
